refactor: 베팅값 입력 범위 변경

+++ src/main/java/blackjack/dto/PlayerBettingRequest.java
@@ -2,7 +2,7 @@ package blackjack.dto;
 
 public record PlayerBettingRequest(
     String playerNickname,
-    int amount
+    long amount
 ) {
 
     public static PlayerBettingRequest of(String playerNickname, String rawAmount) {
